fix up insertion of fragments

+++ src/particleSpatialDistribution.h
@@ -35,7 +35,7 @@ class TriMeshContacts;
 class ParticleSpatialDistribution
 {
  public:
-  ParticleSpatialDistribution(RanPark *rp, int overlap, int maxattempt);
+  ParticleSpatialDistribution(RanPark *rp, double overlap, int maxattempt);
   ~ParticleSpatialDistribution();
 
   bool isPointInSphere(const std::vector<double> &center, double radius, const std::vector<double> &x, double *dir=NULL, double *dist=NULL);
@@ -66,7 +66,7 @@ class ParticleSpatialDistribution
 
  private:
   RanPark * RNG;
-  int max_overlap;
+  double max_overlap;
   int maxattempt;
 };
